requirements.txt
mistralai
pikepdf
python-docx
python-pptx
python-dotenv
